The input text file is read and chunked. The audiobook project name is read from the text element Added a button to cancel the audiobook project creation, but it's not active yet Added a button to create the xml project file. All chunks are written to the XML file.
TODOs: the chunk size needs to be configurable. Currently it's hardcoded to 2800 characters
the Cancel button should lead back to the welcome screen

There appears to be a bug where the area where the FAB menu button is located is clickable, even tho the FAB button itself is hidden. Need to find a way to prevent it from being clickable.