Saving a new empty project now works via the save button As a test, the Start processing button deletes the current project. Database now contains a threadpool Database uses the threadpool to perform queries Inserting multiple items with foreign keys at the same time DOESN'T WORK in Room - apparently this is a bug from 2018 which they've decided to reject as being out of scope with their idea... Inserts must be done entity by entity, synchronously in a thread. Deletes however work automatically, and don't need to be on a separate thread... What the hell, Google? We are now using the (hopefully) correct way to get the navigation controller inside the fragments (NavHostFragment.findNavController(this))
I've added a boolean to the bundle when we navigate to the setup fragment, to indicate if we are creating a brand new project. Perhaps a better way would be to have a whole separate fragment for this?

Finally fixed the idiotic automatic code rearranging
Removed the executor service from MyAudiobookCanvasApplication and from the repository. I should probably get rid of the custom application class entirely at some point.

I've also hit a weird bug where the database isn't updated if I tap save and then try to delete the entry and navigate up...