Updated the TextPreparationFragment - it now has an interactive loading spinner
Added a new TextFileRepository
It uses a background thread, in which it opens the text file, chunks it, and sanitises the chunks and returns them as an array list.

Added a new ViewModel to process text files
It calls the TextFileRepository to obtain text chunks via a listener. It can also store the chunks in our database, and attatch them to the project_id

Added new DAO and relation for a Project with TextBlocks (I should really stop calling them chunks in the code)

Modified the AudiobookRepository to write the text blocks to our database
Added a FileOperationListener interface for callbacks from the file repository operations

Updated the database with the new DAO

Added a new drawable for the progress spinny bar thing